Added SSL network support
SSL sockets return CC_ECONNRESET on a read if a dirty shutdown was
caused by the server (i.e. close_notify was not sent).

+++ utility.c
@@ -366,7 +366,7 @@ uint32_t utf8next(const char *utf8, const char **next) {
     codepoint = (unsigned char) *utf8 & (0xff >> bytesInCode);
 
     /* Obtain continuation bytes */
-    for (size_t i = 1; i < bytesInCode; ++i) {
+    for (int i = 1; i < bytesInCode; ++i) {
         if ((utf8[i] & 0xC0) != 0x80) /* Invalid continuation byte (note this handles a terminating NUL just fine) */
             return error;
